requirements.txt
## 📦 Requirements
streamlit
requests
pandas
beautifulsoup4
googlesearch-python
re
json